Add BackButton to NavBar
