+++ next/src/components/layouts/PageLayout.tsx
@@ -48,6 +48,7 @@ const PageLayout = ({
         ) : (
           <NavBar
             hideNavMenu={hideNavMenu}
+            backButtonHidden={hideBackButton}
             desktopNavbarRef={desktopNavbarRef}
             mobileNavbarRef={mobileNavbarRef}
           />
