+++ next/src/components/segments/NavBar/AuthNavBar.tsx
@@ -1,10 +1,7 @@
-import { Button } from '@bratislava/component-library'
-import { useRouter } from 'next/router'
-import { useTranslation } from 'next-i18next/pages'
 import { RefObject } from 'react'
 
-import Icon from '@/src/components/icon-components/Icon'
 import SectionContainer from '@/src/components/layouts/SectionContainer'
+import BackButton from '@/src/components/segments/NavBar/BackButton'
 import OAuthLogo from '@/src/components/segments/OAuthLogo/OAuthLogo'
 import { AlertBanner } from '@/src/components/simple-components/AlertBanner'
 import Brand from '@/src/components/simple-components/Brand'
@@ -16,25 +13,6 @@ type Props = {
   desktopNavbarRef: RefObject<HTMLDivElement | null>
   mobileNavbarRef: RefObject<HTMLDivElement | null>
   className?: string
-}
-
-const BackButton = () => {
-  const { t } = useTranslation('account')
-  const router = useRouter()
-
-  return (
-    <>
-      <Button
-        variant="icon-wrapped-negative-margin"
-        size="large"
-        icon={<Icon name="arrow-left" />}
-        aria-label={t('BackButton.aria')}
-        onPress={() => router.back()}
-        className="max-lg:mx-1"
-      />
-      <div className="mx-6 h-6 border-r max-lg:hidden" aria-hidden />
-    </>
-  )
 }
 
 /**
