+++ next/src/components/segments/NavBar/NavBar.tsx
@@ -10,12 +10,13 @@ type Props = {
   desktopNavbarRef: RefObject<HTMLDivElement | null>
   mobileNavbarRef: RefObject<HTMLDivElement | null>
   hideNavMenu?: boolean
+  backButtonHidden?: boolean
 }
 
 /**
  * Figma: https://www.figma.com/design/17wbd0MDQcMW9NbXl6UPs8/DS--Component-library?node-id=19549-21360&t=EGiWvvrAjJLDEfQk-4
  */
-const NavBar = ({ desktopNavbarRef, mobileNavbarRef, hideNavMenu }: Props) => {
+const NavBar = ({ desktopNavbarRef, mobileNavbarRef, hideNavMenu, backButtonHidden }: Props) => {
   const alertRef = useRef<HTMLDivElement>(null)
   const { height } = useResizeObserver({ ref: alertRef as React.RefObject<HTMLElement> })
 
@@ -24,13 +25,17 @@ const NavBar = ({ desktopNavbarRef, mobileNavbarRef, hideNavMenu }: Props) => {
       <div className="hidden w-full bg-background-passive-base lg:block" ref={desktopNavbarRef}>
         <AlertBanner />
         <div className="relative w-full">
-          <NavBarHeader />
+          <NavBarHeader backButtonHidden={backButtonHidden} />
           {!hideNavMenu && <NavMenu />}
         </div>
       </div>
       <div style={{ height }} aria-hidden className="hidden lg:block" />
 
-      <MobileNavBar mobileNavbarRef={mobileNavbarRef} className="lg:hidden" />
+      <MobileNavBar
+        mobileNavbarRef={mobileNavbarRef}
+        backButtonHidden={backButtonHidden}
+        className="lg:hidden"
+      />
     </>
   )
 }
